Fixing issues with display of GC data
On the hub, the GC portion of the order display was showing an error.
In the cart_service.js, the variable name did not match what was used
in checkout_controller.js. This was causing giftCertRemain to be
unexpectedly undefined.

This change adds an additional property to sales,
“gift_cert_available”, and adds proper docs on what each one of these
properties is used for in cart_service.js.